# components/WebServer/public/apidoc/swagger.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: LinSTT
  description: A simple way to test LinSTT API, and to learn how to use API
schemes:
  - http
host: localhost:8000
basePath: /
paths:
...
  /service/{serviceId}:
    post:
      tags:
        - "Service Manager APIs"
      summary: Create the service by serviceId
      produces:
      - "application/json"
      parameters: 
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      - name: "replicas"
        in: "path"
        description: "number of replicas"
        required: true
        schema:
          type: "string"
      - name: "tag"
        in: "path"
        description: "Service mode"
        required: true
        schema:
          type: "string"
      - name: "languageModel"
        in: "path"
        description: "Language model to use"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Service 'serviceId' is successfully created
          schema:
            type: "object"
            properties:
              status:
                type: "string"
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
    put:
      tags:
        - "Service Manager APIs"
      summary: Update the service by serviceId
      produces:
      - "application/json"
      parameters: 
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      - name: "replicas"
        in: "path"
        description: "number of replicas"
        required: true
        schema:
          type: "string"
      - name: "tag"
        in: "path"
        description: "Service mode"
        required: true
        schema:
          type: "string"
      - name: "languageModel"
        in: "path"
        description: "Language model to use"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Service 'serviceId' is successfully update
          schema:
            type: "object"
            properties:
              status:
                type: "string"
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
    delete:
      tags:
        - "Service Manager APIs"
      summary: delete the service by serviceId
      produces:
      - "application/json"
      parameters: 
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Service 'serviceId' is successfully removed
          schema:
            type: "object"
            properties:
              status:
                type: "string"
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
    get:
      tags:
        - "Service Manager APIs"
      summary: get the service information by serviceId
      produces:
      - "application/json"
      parameters:
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Json document
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
  /service/{serviceId}/replicas:
    get:
      tags:
        - "Service Manager APIs"
      summary: get the number of replicas by serviceId
      produces:
      - "application/json"
      parameters:
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Json document
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
  /service/{serviceId}/mode:
    get:
      tags:
        - "Service Manager APIs"
      summary: get the decoding mode by serviceId
      produces:
      - "application/json"
      parameters:
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Json document
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
...
  /service/{serviceId}/start:
    post:
      tags:
        - "Service Runtime APIs"
      summary: start service by serviceId
      produces:
      - "application/json"
      parameters:
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Service 'serviceId' is successfully started
          schema:
            type: "object"
            properties:
              status:
                type: "string"
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
  /service/{serviceId}/stop:
    post:
      tags:
        - "Service Runtime APIs"
      summary: stop service by serviceId
      produces:
      - "application/json"
      parameters:
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Service 'serviceId' is successfully stopped
          schema:
            type: "object"
            properties:
              status:
                type: "string"
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
  /service/{serviceId}/scale/{replicas}:
    post:
      tags:
        - "Service Runtime APIs"
      summary: start service by serviceId
      produces:
      - "application/json"
      parameters:
      - name: "serviceId"
        in: "path"
        description: "Service name"
        required: true
        schema:
          type: "string"
      - name: "replicas"
        in: "path"
        description: "Number of replicas"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: Service 'serviceId' is successfully scaled up/down
          schema:
            type: "object"
            properties:
              status:
                type: "string"
        400:
          description: Error message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
...